Code Update :- UIActivityViewController explained
Sharing things using iOS uses a standard, powerful component that other apps can plug into. As a result, it should be your first port of call when adding sharing to an app.
This component is called UIActivityViewController: you tell it what kind of data you want to share, and it figures out how best to share it.

Link :- https://www.hackingwithswift.com/read/3/2/uiactivityviewcontroller-explained