Code Update : Pick a word, any word: UIAlertController
This game will prompt the user to enter a word that can be made from the eight-letter prompt word. For example, if the eight-letter word is "agencies", the user could enter "cease." We're going to solve this with UIAlertController, because it's a nice fit, and also gives me the chance to introduce some new teaching. I'm all about ulterior motives!

https://www.hackingwithswift.com/read/5/3/pick-a-word-any-word-uialertcontroller